ENH: Added mant_dig attribute to finfo (#22333)
The normal attribute for getting the mantissa is
nmant, but this is always off by 1. However, other
packages (like sympy) require the nmant attribute,
so we would end up breaking code downstream if we
changed it. Instead, I introduced a new attribute
called mant_dig, as member Sebastian Berg suggested.
In the documentation for finfo, I added a small
disclaimer saying that nmant should not be used
anymore. Since nmant is going to have to be
deprecated eventually, I also added a small
disclaimer that nmant is inaccurate, set for
depreciation at some point in the future,
and that mant_dig should be used from here on out.