TST, BLD: Fix failing aarch64 wheel builds.
The aarch64 wheel build tests are failing with OOM. The new test for
complex128 dot for huge vectors is responsible as the useable memory
is incorrectly determined and the check for sufficient memory fails.
The fix here is to define the `NPY_AVAILABLE_MEM="4 GB"` environment
variable before the test call in `cibw_test_command.sh`.