ENH: Expose `ufunc.resolve_dtypes` and strided loop access
The API here designed is "future" in the sense that it implementes
NEP 50 and exposes loop specializations as per NEP 43 which is barely
used by NumPy itself at this point.

Due to the fact that NEP 50 is not implemented (or rather finalized)
this API is not ideal and creates dummy-arrays internally so that
it is not much faster than if the user created dummy arrays to
probe the correct result.