MAINT: Doing float comparisons in DivisionSafeDomain only makes sense for floats
Sure, it barely makes sense there as well, but at least this way we can
bypass it and only do the "nonzero" check, which also makes sense for
timedelta64.

And we need timedelta64 to work, because `np.nanmedian` uses the Masked
median internally, and that ends up here :tada: (sarcasm).