MAINT: (array-coercion) Silence invalid read warning in some gcc versions
The warnings are harmless and seem to be caused by the "hint" that the
shape cannot be larger than that area.  But some compiler versions seem
to check that the function call is always correct based on the signature.

That probably makes sense, so just remove that "size hint" which is maybe
even technically incorrect (the compiler is not allowed to read more as
an optimization).