MAINT: remove macOS specific long double handling in numpyconfig.h
This was put in place for universal builds (the old kind, PPC/Intel),
and then extended for arm64 support. It was only needed when building
for two architectures in a single build.

We no longer support i386/PPC/universal, and for producing universal2
wheels for those users that want that, the way to do it is to
take a x86-64 wheel and an arm64 wheel and fuse those with the
`delocate-fuse` utility from `delocate`.

Hence this code is no longer needed.