API: (cython) remove `long_t` and `ulong_t`
They are both very confusing aliases.  Unfortunately, I did not find
a way to give a more informative cython compilation error.
Thus, I pasted the expected error message in the hope it will be
easy to google.

The release notes are long and maybe confusing.  I really want to
prepare us for switching the default integer to (probably) `intp`
and this could be rather disprutpive for Cython modules if, so
I thought it might be good to hint towards that, but maybe that is
too much?