ENH: Implement matmul using the nuclear options
This uses the `axes` argument.  Arguably the most correct version since
we use the full ufunc machinery, so we can't mess up with random conversions
(which the test suite actually did notice, at least for an array subclass).

OTOH, for now the errors are ridiculously unhelpful, some of which could
be fixed in the gufunc code (at least made _better_).