BUG: Ensure string aliases `"int0"`, etc. remain valid for now
int0 and uint0 were accidentally dropped, the others just added as
a test.
We did successfully remove many Numeric types before, so these could
probably be deprecated (it is a bit more annoying to do).

These could probably just be removed, scikit-learn only noticed it in
a single test (scipy probably not at all).  But maybe not in 1.24