DEP: Finalize the comparison FutureWarning/DeprecationWarning
This finalize the deprecation warning, there are a couple of corner
cases that we don't get as nicely as we could.  E.g. we (for now?)
manually create the all-false or all-true result array with a bit
clunky subclass support (subclasses can always override `==` and
`!=` though).
We also keep some (existing) behavior for 0-D objects where we
just return `NotImplemented`, which means the result should end
up with Python booleans (which is probably just as well).