ENH: Speedup masked array creation when mask=None (#22725)
This PR Closes gh-17046.

The problem was that when calling mask=None, the array creation took seconds compared to the microseconds needed when calling mask=False.

Using `mask=None` is a bit dubious, since it has a different meaning from the default `mask=nomask`, but the speed trap is so hard to find, that it seems pragmatic to support it.  OTOH, it also would seem fine to deprecate the whole path (or maybe see if the path can be sped up so that the speed difference isn't forbidding eough to bother).