API: Hide exceptions from the main namespace
I wasn't sure if we should already start deprecating the exceptions
so opted to follow up with only hiding them from `__dir__()` but
still having them in `__all__` and available.

This also changes their module to `numpy.exceptions`, which matters
because that is how they will be pickled (it would not be possible
to unpickle such an exception in an older NumPy version).

Due to pickling, we could put off changing the module.