BUG: fix unexpected return of np.pad with mode=wrap (#22575)
np.pad with mode="wrap" returns unexpected result that original data is not strictly looped in padding. This may happen in some occassions when padding widths in the same dimension are unbalanced (see added testcase in test_arraypad.py and the related issue). The reason is the function pad makes iterative calls of _set_wrap_both() in the above situation, yet period for padding is not correctly computed in each iteration.

The bug is fixed by guaranteeing that period is always a multiple of original data size, and also be the possible maximum for computation efficiency.

Closes #22464

Co-authored-by: Lars Grüter <lagru+github@mailbox.org>