BUG: Do not use getdata() in np.ma.masked_invalid
This is the minimal solution to fix gh-22826 with as little change
as possible.
We should fix `getdata()` but I don't want to do that in a bug-fix
release really.

IMO the alternative is to revert gh-22046 which would also revert
the behavior noticed in gh-22720  (which seems less harmful though).

Closes gh-22826