BUG: Fortify string casts against floating point warnings
This removes the check for floating point warnings, which is enough
in practice.  (In principle ufuncs or structured dtypes can chain
casts in a way that causes us to check anyway.)

It also checks for isfinite in the scalar repr code so the warnings
shouldn't be set to begin with.

Closes gh-22843