ENH: Add __array_function__ implementation to MaskedArray
Previously, calling np.stack(ma, a) with a masked and normal array
would stack the nd-data from the items, but wrap it in a MaskedArray
class without implementing the mask. This implementation intercepts
the np.func calls and checks whether there is a compatible masked
version that we can use instead. Now, the mask gets propagated with
the stack as expected.