MAINT: Add additional information to missing scalar AttributeError
This is a followup on gh-22607 which removed them.  Since it appears some
users missed the DeprecationWarning entirely, it may help them to
include the old information as an attribute error.

An example is:
```
In [1]: np.int

AttributeError: module 'numpy' has no attribute 'int'.
`np.int` was a deprecated alias for the builtin `int`. To avoid this error in existing code, use `int` by itself.
Doing this will not modify any behavior and is safe. When replacing `np.int`, you may wish to use e.g. `np.int64`
or `np.int32` to specify the precision. If you wish to review your current use, check the release note link for
additional information.
The aliases was originally deprecated in NumPy 1.20; for more details and guidance see the original release note
at:
    https://numpy.org/devdocs/release/1.20.0-notes.html#deprecations
```

Yes, that is very verbose...

your changes. Lines starting