MAINT: Move export for scipy arm64 helper into main module
This is a follow up to gh-22679 which addressed gh-22673.

The main thing is that we want the functions to be available after
importing NumPy, so they need to be part of multiarray.
However, `npymath` is a static library, so the symbols are not really
exported there.  The former PR did actually work in practice but this
seems like it is technically the right place?

For some reason, I had to add nextafter to be able to do:

    from scipy.spatial.distance import euclidean

with the SciPy 1.9.3 wheels.  SciPy test collection works with this for
the 1.9.3 wheel, so this should be all the symbols hopefully.