BUG: Ensure correct loop order in sin, cos, and arctan2
These were incorrect afer being vectorized.  The commit additional
tests these (not arctan2 admittedly) and adds a check to generate_umath
to make it a bit less likely that future additions add this type of thing.

Note that the check allows duplicated loops so long they are correctly
ordered the *first* time.  This makes results correct, but duplicated
loops are not nice anyways and it would be nice to remove them.

We could drop them manually in hindsight even?  In any case, that should
not be backported, so it is not includedhere.

Closes gh-22984