DEP: Finalize the non-sequence stacking deprecation
The `__array_function__` API currently will exhaust iterators so we
cannot accept sequences reasonably.  Checking for `__getitem__` is presumably
enough to reject that (and was what the deprecation used).

Future changes could allow this again, although it is not a useful API
anyway, since we have to materialize the iterable in any case.