BUG: Fixup f2py's handling a very little bit
This clears the error holding only to the type.  Since in the other
path the errmessage seemed completely uninitialized, I opted to just
ignore it entirely and keep the old error.

I could fathom to use error chaining here, but overall, I am not even
sure that chaining makes even sense for these errors.  This fix is
meant to be minimal (the second one, I just noticed randomly), it
does not make this code clean.