ENH: Allow trivial pickling of user DType (classes)
This also adds a `_legac` attribute, I am happy to rename it, but
I suspect having something like it is useful.

Arguably, the best solution would be to give our DTypes a working
module+name, but given that we are not there, this seems easy.
We could probably check for `__reduce__`, but I am not certain that
wouldn't pre-empt successfully already, so this just restores the
default for now.