MAINT: Use strong references/copies for sorting buffer
I accidentally included a start for a cleanup in the other PR which
was incorrect because we actually sorted weak references.

Sorting weak references buffer is correct, but seems a bit trickier
to reason about and also potentially to generalize.

This makes sure we have strong references everywhere and fixes the
issue seen by pandas.  Also adds a (slightly complex) test to cover
both the sort and argsort path.