MAINT: Fix failing gitpod build.
The gitpod builds have begun failing, this PR updates the
`build-push-action` command to version 4 to check if that fixes the
problem. Note the the build only runs on merge, so the PR tests won't
help, but a version update is probably a good idea anyway. See gh-23320
for more information.