ENH: improve dlpack protocol detection
Currently, dlpack protocol detection is done by checking the existence
of `__dlpack__` method on the object's type. This is not enough to
detect the dlpack protocol support, because the method can be defined on
the object itself, not on the object's type.
This patch improves the detection by checking the existence of
`__dlpack__` method on the object itself.

Signed-off-by: Gigon Bae <gigony@gmail.com>