BLD: Add support for NPY_TARGET_VERSION macro
This is a way for downstream users to specify which NumPy version
they wish to be compaible with.

Note that we provide a conservative default here (because almost nobody
actually uses new API as they would lose backwards compatibility).

Initially I had thought we should just redefine it so that the target
version uses the same scheme as the Python hex version (and limited API),
but since we have `NPY_1_15_API_VERSION` defines, use those...

This commit does not include any actual use of this!