MAINT: Add a proper implementation for structured zerofill
This reorganizes the generic traversal functions for structured
dtypes a bit (before it wasn't quite generic).

Then uses that for zerofilling.  We could get the two a bit closer
by also supporting `func==NULL` explicitly for clearing.
(I have not includes this here.)

The old `FillObjectArray` is still in use now, this is not ideal
and the new approach should be duplicated to add an "emptyfill"
(same semantics, normally just memset to 0, but for objects we
place an explicit `None` object).

This is a necessary follow-up to gh-23591.