BUG: Fix masked array raveling when `order="A"` or `order="K"`
This transitively fixes gh-22912.  I had alooked a bit into whether
it is worthwhile to preserve the mask order, but TBH, we seem to not
do so in so many places, that I don't think it really is worthwhile.

Applying `order="K"` or `order="A"` to the data and mask separately
is an big bug though.

Closes gh-22912