ENH: Restore TypeError cleanup in array function dispatching
When the dispathcer raises a TypeError and it starts with the dispatchers
name (or actually __qualname__ not that it normally matters), then it is
nicer for users if we just raise a new error with the public symbol name.

Python does not seem to normalize exception and goes down the unicode path,
but I assume that e.g. PyPy may not do that.  And there might be other
weirder reason why we go down the full path.  I have manually tested it
by forcing Normalization.

Closes gh-23029