MAINT: Reorganize the way windowing functions ensure float64 result
This roughly changes things so that we ensure a float64 working
values up-front.  There is a tiny chance of precision changes if the
input was not float64 or error changes on bad input.

I don't think this should matter in practice, precision changes
(as far as I can tell) should happen rather the other way around.

Since float64 has 53bits mantissa, I think the arange should give
the correct result reliably for any sensible inputs.

There is an argument to be made that the windowing functions could
return float32 for float32 input, but I somewhat think this is OK
and users can be expected to just cast manually after the fact.

The result type is tested, but this ensures the tests pass also
when enabling weak promotion.