MAINT: Remove gisnan, gisinf, and gisfinite from testing code
These were introduced many years ago when ufuncs were buggy and
could return NotImplemented sometimes.  This has been fixed for
many years, though.

I suspect the errstate for `isfinite` is not required so removed it.
It was a 12+ year old work-around for warnings that really shouldn't
happen to begin with.  (The commit mentions `np.isinf(np.inf)` giving
a warning, which doesn't make sense, I think.)