MAINT: Do not use copyswap as part of indexing
This is in a slower effort to just avoid copyswap.  It is very
slow for structured dtypes and otherwise, I would prefer a new
fast "single element copy" functionality (or as part of the
normal strided loops, a single element special case?)

To some degree, it could be nice to have a copy single element
functionality more specific, but we probably need the `get_func()`
pattern for cases like here, and at that point this doesn't add
much boilerplate additionally?

This fixes new DTypes that don't need a `copyswap` right now
(and I would like them not needing it).