ENH: allow int sequences as shape arguments in numpy.memmap
The constructor for memmap objects initially required the shape argument
to be a tuple, where non-tuple shape arguments would be made into a
single member tuple. This behaviour was intended to allow an integer
shape arguments. However this caused list or numpy.array shape arguments
to raise errors within the constructor. Modifications made to the
shape argument handling within the constructor now allow for
integer sequences to be used for shape. A test was added and
constructor docstring was updated to account for these changes.