BUG: Fix segfault in timedelta related to int/float
gentype having int/float that goes to array, and array having one
that goes to the scalar doesn't make sense.  So just solve the issue
by making the gentype version always an error.

(Could put it on the datetime, but it seems not useful.
Simply not implementing it isn't great, because `int()` seems to try
to interpret things as a buffer somehow.)