BUG: Fix median and quantile NaT handling
Note that this doesn't mean that rounding is correct at least for
quantiles, so there is some dubious about it being a good idea to
use this.

But, it does fix the issue, and I the `copyto` solution seems rather
good to me, the only thing that isn't ideal is the `supports_nan`
definition itself.

Closes gh-20376