CI: bump nightly upload frequency to twice a week
With Nathan working on dtypes I occasionally triggered wheels on
CI manually and I think we have a decent amount of PRs merged every
day so that twice a week seems fine.

Time chosen to be a mix of what we had on cirrus and here (this
aligns it, but I am happy to adjust both).

[skip ci]