DOC: Note that f2py isn't consiered safe
Not mentioning distutils explicitly, because it is deprecated anyway.
This is basically just because occasionally get those "this is an unsafe
regex", etc.  Also, I would not be surprised if f2py isn't full of actual
bugs that would allow escalation if you use it on untrusted input.

I don't think it really needs pointing out, but adding it so we have
something to point at if it comes up again.