ENH: Introduce new functions that return both the standard deviation and the mean: mean_std and the variance and the mean: mean_var.
IN the current setup the mean is often calculated twice, this can be prevented if the functions calculating the variance of the standard deviation also return the mean,

Fix typo's, minor problems and add tests to increase coverage.

Note:
fromnumeric.py is littered with constructions like:

```Python
    if type(a) is not mu.ndarray:
        try:
            var = a.var
        except AttributeError:
            pass
        else:
            return var(axis=axis, dtype=dtype, out=out, ddof=ddof, **kwargs)
```
This looks like untested defensive coding, and because the data object is not passed to the method it is disfunctional. I copied it, but I think it can go. Which would be my preferred way of improving the code coverage here.