API: Remove `PyUFunc_` APIs `GetPyValues`, `checkfperr`, and `handlefperr`
We could introduce a new to combine these (mainly the GetPyValues and
handlefperr) and we could do so in a way that is backwards compatible:
Combine both functions into one small helper.

There is no known user, but if someone comes up even restoring them
should be possible.

These functions are threaded through our code to deal with error handling
but create an unnecessary `errobj` that may have made some sense when
you wish to raise errors from a worker/inner-loop.
However, we never do that, and there shouldn't a be reason for downstream
to do that either:  The current split in functions just makes it hard
to improve the `errstate` handling (because this API will not remain useful
internally).