BUG: Make errstate decorator compatible with threading
We need to store the token per-thread, which doesn't work if one
instance of errstate is shared between threads (and we store it on
the instance).  `ContextDecorator` has a mechanism for this, but
it is not public.

So, this just implements it the full manual style.

I additionally add a new note on async-safety and the missing test
(yes, fails on 1.25).

Closes gh-24013