BUG: Multiple or Divides using SIMD without a full vector can cause divide-by-zero
The case occurs whenever a multiple or divide occurs (`array * scalar`) using SIMD and not a full vector can be used.  The remaining lanes get zeroed, which can cause divide-by-zero or invalid to be raised.

Xcode 14.3 includes an updated clang that supports FP Strict for arm64, which does not have any issues.

Previous versions of Xcode / clang have incorrect codegen for this case.