MAINT: Remove ability to enter errstate twice (sequentially)
That was a bit of a slip.  The issue why I allowed it was that
it "fixed" the decorator version, but it was not the correct fix.

Considering that things like `warnings.catch_warnings()` do not allow
this either, this seems better.
I also don't think it was technically thread-safe.