MAINT: Allow all options again (because subarrays need them) and fix test
If we use a format string, just ignore all other options and grab
the global ones.  I am not 100% sure if there might be a weird path
where things get lost on the way for structured arrays or so,
this change is new, I so I think its fine eithe  way.

(and I suspect it is actually correct)