BUG: Fix empty structured array dtype alignment
This change ensures that dtypes are given an alignment of at least 1.
Prior to this change, dtypes of empty structured arrays with `align` set
to True would have an alignment of 0.

For instance, `np.dtype([], align=True).alignment` returns 0. Empty
structured array dtypes should have an alignment of 1, meaning that they
do not need padding. An alignment of 0 is problematic because the
alignment is used to determine offsets, which involves dividing by the
alignment. If the alignment is 0, then a division by zero occurs.

The following code raises a ZeroDivisionError when converting the dtype
to a string representation:

```python
repr(np.dtype([('f0', [])], align=True))
```

The following code crashes Python because a division by zero occurs
in the C code:

```python
np.dtype({'names':['f0'], 'formats':[[]], 'offsets': [0]}, align=True)
```