BUG: Only replace dtype temporarily if dimensions changed
For strings this is (currently) arguably not correct... Now of course
strings shouldn't be changed to S0 probably, but that is a different
matter.

It seems there is something odd (possibly I just kept it as a corner
case) for coercion from object arrays.

Tests should cover the paths I think (lets see what CI says) but
I couldn't reliably trigger failures for all paths.

Closes gh-24043