update the dispatchers according to suggestions by Marten van Kerkwijk:
(https://github.com/numpy/numpy/pull/24126#discussion_r1254314302) The dispatcher returns all arguments that can, in principle, contain something that is an array and hence that, if not a regular ndarray, can allow the function to be dealt with another package (say, dask). Since mean can be an array, it should be added (most similar to where).