MAINT: Update meson.build files from main branch
In addition to the meson.build files in numpy, the submodules
`numpy/core/src/npysort/x86-simd-sort` and `numpy/core/src/umath/svml`
have also been updated to get the meson build changes in them.

The include files have not been updated, as the main change seems to
be removal of some of them in line with the changes in 2.0.0. Likewise
the ABI and API numbers in `numpy/core/meson.build` have been kept in
line with 1.25.x, but the dependencies there should be checked by
someone more familiar with the new build infrastructure.