BUG: fix invalid function pointer conversion error
`f2py_init_func` is defined in `fortranobject.h` as:

    typedef void (*f2py_init_func)(int *, npy_intp *,
                  f2py_set_data_func, int *);

`f90mod_rules.py` attempts to assign functions with the following
prototype to `func` of `struct FortranDataDef`:

    void(*)(int*, int*, void(*)(char*, int*), int*)

This causes the `test_module_doc.py` test to fail when `npy_intp` is
`long` and clang 16 is used as the compiler because the assignment is
identified as an invalid function pointer conversion, which clang 16
treats as an error by default.

Using `npy_intp` instead of `int` fixes the error.