Remove unnecessary legacy check and longdouble quoting support
Undoing the other stuff, basically means that quoting was now used
where it is unclear that it should be (in array printing/repr) but
not where it definitely should be (in void repr).

So there is just no point, we basically decided to not support it
for now and consider it a bug to be fixed later.