BUG: random: Fix generation of nan by dirichlet.
Don't call the C function random_beta() with both parameters
`a` and `b` set to 0.  In the case where this would occur, we
know that the remaining values in the random vector being
generated must be 0, so can break out of the loop early.

After this change, when alpha is all zero, the random variates
will also be all zero.

Closes gh-24210.