BUG: Fix cblas detection for the wheel builds
This is a mess and doesn't attempt to fix the issues around gh-24200, but
it should fix the issue of our wheels not using CBLAS.

The suffix may be directly injected into the flags to get the define and
there is also a tangle between `HAVE_CBLAS` and `HAVE_LAPACK` and we only
probe the first.

I tried this locally by checking that *if* the define exist redefining
it to nonesense gives the slow version.
(I cannot test the full thing without a prefixed blas locally, because
setting the prefix will not avoid blas use in LAPACK meaning you don't
use CBLAS, but the blas symbols used by lapack are missing)