MAINT: exclude min, max and round from `np.__all__`
This is a workaround for breakage in downstream packages that
do `from numpy import *`, see gh-24229.

Note that there are other builtins that are contained in `__all__`:

```
>>> for s in dir(builtins):
...     if s in np.__all__:
...         print(s)
...
all
any
divmod
sum
```

Those were already there before the change in 1.25.0, and can stay.
The downstream code should be fixed before the numpy 2.0 release.

Closes gh-24229