BUG: Factor out slow `getenv` call used for memory policy warning
Using `getenv` regularly is probably not great anyway, but it seems
very slow on windows which leads to a large overhead for every array
deallocation here.

Refactor it out to only check on first import and add helper because
the tests are set up slightly differently.
(Manually checked that the startup works, tests run with policy
set to 1, not modifying it and passing.)