BUG: random: Fix generation of nan by beta.
The implementation of Johnk's algorithm for beta(a, b) could generate nan
values if both a and b were extremely small (i.e. subnormal or a small
multiple of the smallest normal double precision float).

The fix is to handle variate generation in this case by noting that when
both a and b are extremely small, the probability of generating a double
precision value that is not either 0 or 1 is also extremely small.  In
particular, if a and b are less than 3e-103, the probability of generating
a (double precision) value that is not 0 or 1 is less than approximately
1e-100.  So instead of using Johnk's algorithm in this extreme case, we
can generate the values 0 or 1 as Bernoulli trials, with the probability
of 1 being a/(a + b).

Closes gh-24266.